samadhan/backend/requirements.txt
google-genai>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
groq>=0.4.0
